+++ pyproject.toml
@@ -22,7 +22,7 @@ requires = ["poetry-core"]
 build-backend = "poetry.core.masonry.api"
 
 [tool.poetry.scripts]
-competitive-verifier = "competitive_verifier.console.app:main"
+competitive-verifier = "competitive_verifier.app:main"
 
 [tool.poetry.dependencies]
 python = ">=3.9,<4.0"
@@ -64,11 +64,19 @@ test = "pytest"
 test-quick = "pytest -m 'not integration'"
 test-integration = "pytest -m integration"
 
-"dev:verify" = { cmd = "competitive-verifier verify -o .competitive-verifier/tmp/result.json", env = { COMPETITIVE_VERIFY_FILES_PATH = "examples/external/verify_files.json" } }
-"dev:docs" = { cmd = "competitive-verifier docs .competitive-verifier/tmp/result.json", env = { COMPETITIVE_VERIFY_FILES_PATH = "examples/external/verify_files.json" } }
 
 "jekyll:serve" = { cmd = "bundle exec jekyll serve --incremental --livereload", cwd = ".competitive-verifier/_jekyll" }
 "jekyll:init" = { shell = "bundle config set --local path '.vendor/bundle'; bundle install", cwd = ".competitive-verifier/_jekyll" }
+
+
+[tool.poe.tasks."dev:verify"]
+script = "competitive_verifier.verify.main:main(['--output', '.competitive-verifier/tmp/result.json'])"
+env = { COMPETITIVE_VERIFY_FILES_PATH = "examples/external/verify_files.json" }
+
+
+[tool.poe.tasks."dev:docs"]
+script = "competitive_verifier.docs.main:main(['.competitive-verifier/tmp/result.json'])"
+env = { COMPETITIVE_VERIFY_FILES_PATH = "examples/external/verify_files.json" }
 
 [tool.poe.tasks.serve]
 [[tool.poe.tasks.serve.sequence]]
